requirements.txt
streamlit
pyjwt
bcrypt
python-dotenv
pyngrok
nltk
streamlit-option-menu
plotly
textstat
PyPDF2
beautifulsoup4
streamlit-autorefresh
wordcloud
matplotlib
langdetect
pandas
transformers
torch
sentence-transformers
faiss-cpu
accelerate
spacy
networkx
pyvis
bitsandbytes
anthropic